admin organizations

+++ generate-publish-folder.sh
@@ -20,9 +20,16 @@ for dir in */; do
         csproj_file=$(find . -maxdepth 1 -name "*.csproj")
 
         if [ ! -z "$csproj_file" ]; then
-            echo "Publishing project in $dir ..."
-            rm -r ./publish
-            dotnet publish "$csproj_file" -o "./publish"
+            csproj_filename=$(basename "$csproj_file")
+            if [ "$csproj_filename" == "Rackdog.csproj" ]; then
+                echo "Skipping publishing for Rackdog.csproj in $dir."
+            else
+                echo "Publishing project in $dir ..."
+                dotnet nuget locals all --clear
+                dotnet add package rackdog.Rackdog
+                rm -r ./publish
+                dotnet publish "$csproj_file" -o "./publish"
+            fi
         else
             echo "No .csproj file found in $dir, skipping publish."
         fi
